updated ui module to render previews
