+++ .gitignore
@@ -83,3 +83,4 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
+/.idea/inspectionProfiles/Project_Default.xml
