+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="../../../../layout/compose-model-1661973799743.xml" value="0.32222222222222224" />
         <entry key="../../../../layout/compose-model-1662207257062.xml" value="0.33055555555555555" />
+        <entry key="../../../../layout/compose-model-1662247830014.xml" value="0.5" />
       </map>
     </option>
   </component>
