+++ app/build.gradle
@@ -34,7 +34,7 @@ android {
         targetCompatibility JavaVersion.VERSION_11
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = JavaVersion.VERSION_11
     }
     buildFeatures {
         viewBinding true
@@ -69,8 +69,6 @@ dependencies {
     implementation AndroidX.activityCompose
     implementation AndroidX.navigationFragment
 
-    implementation Utils.easyPermissions
-
     implementation Net.retrofit
     implementation Net.okHttpLogging
 
@@ -103,8 +101,6 @@ dependencies {
 
     debugImplementation Net.chuckerDebug
     releaseImplementation Net.chuckerRelease
-
-    debugImplementation Compose.tooling
 
     testImplementation TestLib.junit
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
