+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TrackIt"
         tools:targetApi="31">
-        <service
-            android:name=".MyService"
-            android:enabled="true"
-            android:exported="true"></service>
 
         <activity
             android:name=".MainActivity"
