--- common/ui/build.gradle.kts
@@ -1,51 +1,0 @@
-plugins {
-    id("com.android.library")
-    id("kotlin-android")
-}
-apply(from = "${rootProject.projectDir}/ktlint.gradle")
-
-android {
-    compileSdk = 32
-
-    defaultConfig {
-        minSdk = 23
-        targetSdk = 32
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.3.0"
-
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-
-    implementation(Compose.ui)
-    implementation(Compose.toolingPreview)
-    implementation(Compose.foundation)
-    implementation(Compose.material)
-    implementation(Compose.lifecycleViewModel)
-    implementation(Compose.runtimeLiveData)
-    debugImplementation(Compose.tooling)
-
-}
